fix: CheckBoxRow 위젯파일에서 삭제
회원가입 뷰에만 쓰는 CheckBoxRow 위젯을 AgreementScreen으로 통합함
